+++ ft_type_u.c
@@ -51,7 +51,7 @@ void	ft_case2_u(char **s, size_t l, t_pmts *p, t_prts **n)
 	size_t	var;
 
 	p->prec_value *= -1;
-	if (p->prec_value > l)
+	if (p->prec_value > (int)l)
 	{
 		var = p->prec_value - l;
 		tmp = ft_malloc_sz(var);
