+++ ft_type_x.c
@@ -66,7 +66,7 @@ void	ft_case2_x(char **s, size_t l, t_pmts *p, t_prts **n)
 	size_t	var;
 
 	p->prec_value *= -1;
-	if (p->prec_value > l)
+	if (p->prec_value > (int)l)
 	{
 		if (p->hash)
 			var = p->prec_value - l - 2;
